extracted components into logical files
